Switch target type and parameter type values to integers

+++ lib/Layout/Resolver/TargetType/Subtree.php
@@ -31,7 +31,7 @@ final class Subtree extends TargetType implements ValueObjectProviderInterface
     {
         return [
             new Constraints\NotBlank(),
-            new Constraints\Type(type: 'numeric'),
+            new Constraints\Type(type: 'int'),
             new Constraints\PositiveOrZero(),
             new IbexaConstraints\Location(allowInvalid: true),
         ];
